negamax alpha-beta pruning algo

+++ include/solver.h
@@ -6,7 +6,7 @@ namespace IBN5100 {
     class Solver {
         private:
             uint64_t nodeCount; // track the number of explored nodes
-            int colOrder[7] = {3, 4, 2, 5, 1, 6, 0}; // current priority for the columns
+            uint8_t colOrder[7] = {3, 4, 2, 5, 1, 6, 0}; // current priority for the columns
         
             /**
              * @brief Recursively solve a Connect 4 position using a negamax alpha-beta pruning algorithm.
